[MDH-200] refactor : 점주 모듈 to do 리팩토링(로그인 패키지 변경, 테스트 코드 수정)
