+++ owner/src/main/java/com/mdh/owner/login/presentation/SignUpRequest.java
@@ -1,4 +1,4 @@
-package com.mdh.owner.global.security.controller;
+package com.mdh.owner.login.presentation;
 
 import com.mdh.common.global.util.RegularExpression;
 import com.mdh.common.user.domain.Role;
